pyproject.toml
[project]
name = "prompts-evaluator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.2",
    "google-genai>=1.50.1",
    "openai>=2.8.0",
    "pipecat-ai[azure,daily,google,openai,webrtc]>=0.0.94",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "typer>=0.20.0",
    "uvicorn>=0.38.0",
]
